Migrate 1824 to new structure
